Move remote modules to remote/ folder
