+++ qiskit_alice_bob_provider/ensure_preparation_pass.py
@@ -60,7 +60,7 @@ class EnsurePreparationPass(TransformationPass):
         self._prep = prep_instruction
 
     def run(self, dag: DAGCircuit) -> DAGCircuit:
-        preparations = {'reset', 'initialize', 'state_preparation'}
+        preparations = {'reset', 'initialize', 'remote_state_preparation'}
         for node in dag.topological_nodes():
             if not isinstance(node, DAGInNode):
                 continue
